coded lightning attack

+++ src/Enemies/Enemy.java
@@ -205,7 +205,7 @@ public class Enemy extends MovingImage{
      * Loses a specified amount of health
      * @param x Amount of health lost
      */
-    public void loseHealth(int x) {
+    public void loseHealth(double x) {
     	
     	health -= x;
     	if(health <= 0) {
